# openapi.yaml
openapi: 3.0.0
info:
  title: LeanStock API
  version: 1.0.0
  description: Inventory management system with multi-tenancy, transaction safety and race condition protection
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request
    Unauthorized:
      description: Unauthorized
...
    Conflict:
      description: Conflict occurred
...
  schemas:
...
    AuthRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
paths:
  /auth/register:
    post:
      summary: Register new user
      description: |
        Creates an unverified account. The **first user** in a tenant receives **OWNER**;
        subsequent registrations receive **VIEWER** until an OWNER assigns another role.
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, tenantId]
              properties:
                name:
                  type: string
                  example: "John Doe"
                email:
                  type: string
                  example: "john@example.com"
                password:
                  type: string
                  example: "secret123"
                tenantId:
                  type: string
                  example: 1
      responses:
        "201":
          description: Registered — verification email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Registration successful. Please check your email to verify your account."
                  userId:
                    type: string
                    example: "uuid-here"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/verify-email:
    get:
      summary: Verify email address via token from email link
      tags: [Auth]
      security: []
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
            example: "a3f9c2d1..."
      responses:
        "200":
          description: Email verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Email verified successfully. You can now log in."
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/refresh:
    post:
      summary: Refresh access token (rotates refresh token)
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  example: "uuid-refresh-token"
      responses:
        "200":
          description: New access and refresh tokens
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
        "403":
          description: Invalid or revoked refresh token
  /auth/logout:
    post:
      summary: Logout — revokes refresh token
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  example: "uuid-refresh-token"
      responses:
        "200":
          description: Logged out successfully
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/forgot-password:
    post:
      summary: Request password reset email
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  example: "john@example.com"
      responses:
        "200":
          description: Reset link sent (always 200 to prevent email enumeration)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "If that email exists, a reset link has been sent."
  /auth/reset-password:
    post:
      summary: Reset password using token from email
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token:
                  type: string
                  example: "a3f9c2d1..."
                newPassword:
                  type: string
                  example: "newSecret123"
      responses:
        "200":
          description: Password reset successfully
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/me:
    get:
      summary: Get current authenticated user
      tags: [Auth]
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  email:
                    type: string
                  role:
                    type: string
                    example: CASHIER
                  isVerified:
                    type: boolean
                  active:
                    type: boolean
                  tenantId:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/login:
    post:
      summary: Login user
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        "200":
          description: JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
...